requirements.txt
altgraph==0.17.2
future==0.18.2
openrgb-python==0.2.14
pefile==2021.9.3
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pywin32-ctypes==0.2.0
pywinusb==0.4.2
websocket-client==1.2.3
simpadlib==2.2.1